+++ src/sqlancer/tidb/TiDBErrors.java
@@ -3,9 +3,10 @@ package sqlancer.tidb;
 import java.util.ArrayList;
 import java.util.List;
 
+import sqlancer.SQLErrors;
 import sqlancer.common.query.ExpectedErrors;
 
-public final class TiDBErrors {
+public final class TiDBErrors extends SQLErrors {
 
     private TiDBErrors() {
     }
@@ -68,10 +69,6 @@ public final class TiDBErrors {
         return errors;
     }
 
-    public static void addExpressionErrors(ExpectedErrors errors) {
-        errors.addAll(getExpressionErrors());
-    }
-
     public static List<String> getExpressionHavingErrors() {
         ArrayList<String> errors = new ArrayList<>();
 
@@ -86,25 +83,8 @@ public final class TiDBErrors {
     }
 
     public static List<String> getInsertErrors() {
-        ArrayList<String> errors = new ArrayList<>();
+        List<String> errors = SQLErrors.getInsertErrors();
 
-        errors.add("Duplicate entry");
-        errors.add("cannot be null");
-        errors.add("doesn't have a default value");
-        errors.add("Out of range value");
-        errors.add("Incorrect double value");
-        errors.add("Incorrect float value");
-        errors.add("Incorrect int value");
-        errors.add("Incorrect tinyint value");
-        errors.add("Data truncation");
-        errors.add("Bad Number");
-        errors.add("The value specified for generated column"); // TODO: do not insert data into generated columns
-        errors.add("incorrect utf8 value");
-        errors.add("Data truncation: %s value is out of range in '%s'");
-        errors.add("Incorrect smallint value");
-        errors.add("Incorrect bigint value");
-        errors.add("Incorrect decimal value");
-        errors.add("error parsing regexp");
         errors.add("is not valid for CHARACTER SET");
         errors.add("for function inet_aton");
         errors.add("'Empty pattern is invalid' from regexp");
